+++ ScoreSaber/Core/Data/Models/LeaderboardUploadData.cs
@@ -60,7 +60,11 @@ namespace ScoreSaber.Core.Data.Models {
             data.gameMode = $"Solo{beatmapKey.beatmapCharacteristic.serializedName}";
             data.difficulty = BeatmapDifficultyMethods.DefaultRating(beatmapKey.difficulty);
             data.infoHash = infoHash;
-            data.leaderboardId = levelInfo[2];
+            if(beatmapLevel.hasPrecalculatedData) {
+                data.leaderboardId = "ost_" + beatmapLevel.levelID;
+            } else {
+                data.leaderboardId = levelInfo[2];
+            }
             data.songName = beatmapLevel.songName;
             data.songSubName = beatmapLevel.songSubName;
             data.songAuthorName = beatmapLevel.songAuthorName;
