+++ ScoreSaber/Core/Data/Models/SceneChangeEvent.cs
@@ -1,0 +1,18 @@
+﻿using Newtonsoft.Json.Converters;
+using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ScoreSaber.Core.Data.Models {
+    public class SceneChangeEvent {
+        [JsonProperty("timestamp")]
+        public string Timestamp { get; set; } = string.Empty;
+
+        [JsonProperty("scene")]
+        [JsonConverter(typeof(StringEnumConverter))]
+        public Scene Scene { get; set; }
+    }
+}
