+++ ScoreSaber/Core/Data/Wrappers/LeaderboardInfoMap.cs
@@ -6,12 +6,19 @@ namespace ScoreSaber.Core.Data.Wrappers {
         internal BeatmapLevel beatmapLevel { get; set; }
         internal BeatmapKey beatmapKey { get; set; }
         internal string songHash { get; set; }
+        internal bool isOst { get; set; }
 
         internal LeaderboardInfoMap(LeaderboardInfo leaderboardInfo, BeatmapLevel beatmapLevel, BeatmapKey beatmapKey) {
             this.beatmapLevel = beatmapLevel;
             this.beatmapKey = beatmapKey;
             this.leaderboardInfo = leaderboardInfo;
-            this.songHash = beatmapKey.levelId.Split('_')[2];
+            if (beatmapLevel.hasPrecalculatedData) {
+                this.songHash = "ost_" + beatmapLevel.levelID;
+                this.isOst = true;
+            } else {
+                this.songHash = beatmapKey.levelId.Split('_')[2];
+                this.isOst = false;
+            }
         }
     }
 }
