+++ ScoreSaber/Core/ReplaySystem/Installers/ImberInstaller.cs
@@ -1,5 +1,6 @@
 ﻿using ScoreSaber.Core.ReplaySystem.UI;
 using ScoreSaber.Core.Services;
+using ScoreSaber.Core.Utils;
 using Zenject;
 
 namespace ScoreSaber.Core.ReplaySystem.Installers
@@ -10,13 +11,13 @@ namespace ScoreSaber.Core.ReplaySystem.Installers
 
             if (Plugin.ReplayState.IsPlaybackEnabled && !Plugin.ReplayState.IsLegacyReplay) {
                 Container.Bind<VRControllerAccessor>().AsSingle();
-                Container.Bind<TweeningService>().AsSingle();
+                Container.Bind<TweeningUtils>().AsSingle();
+                Container.BindInterfacesAndSelfTo<DesktopMainImberPanelView>().FromNewComponentAsViewController().AsSingle();
                 Container.BindInterfacesTo<ImberManager>().AsSingle();
                 Container.BindInterfacesAndSelfTo<ImberScrubber>().AsSingle();
                 Container.BindInterfacesAndSelfTo<ImberSpecsReporter>().AsSingle();
                 Container.BindInterfacesAndSelfTo<ImberUIPositionController>().AsSingle();
                 Container.Bind<MainImberPanelView>().FromNewComponentAsViewController().AsSingle();
-                Container.BindInterfacesAndSelfTo<DesktopMainImberPanelView>().FromNewComponentAsViewController().AsSingle();
                 Container.Bind(typeof(ITickable), typeof(SpectateAreaController)).To<SpectateAreaController>().AsSingle();
             }
         }
